Project_1/src/project/Routers/admin.py
from fastapi import APIRouter,Depends,Path,HTTPException
from typing import List,Annotated
from sqlalchemy.orm import Session
from ..db import SessionLocal
from ..models import Todos, User
from starlette import status
from pydantic import BaseModel,Field
from .auth import get_current_user

router = APIRouter(
    prefix="/admin",
    tags=["admin"])
def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()

db_dependency = Annotated[Session,Depends(get_db)]

user_dependency = Annotated[dict,Depends(get_current_user)]



@router.get("/todos",status_code = status.HTTP_200_OK)
async def read_All(user:user_dependency,db:db_dependency):
    if user is None:
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Incorrect username or password")
    return db.query(Todos).all()

@router.delete("/todos/{todo_id}",status_code = status.HTTP_204_NO_CONTENT)
async def delete_todo(user:user_dependency,db:db_dependency,todo_id:int = Path(gt=0)):
    if user is None or user.get('role') != "admin":
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Incorrect username or password")
    todo = db.query(Todos).filter(Todos.id == todo_id).first()
    if todo is None:
        raise HTTPException(status_code=404, detail="Todo not found")
    db.query(Todos).filter(Todos.id == todo_id).delete()
    db.commit()
    return todo

@router.get("/users",status_code = status.HTTP_200_OK)
async def read_All(user:user_dependency,db:db_dependency):
    if user is None or user.get('role') != "admin":
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Incorrect username or password")
    return db.query(User).all()