Project_1/pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = ""
authors = [
    {name = "Anindo Barua",email = "abbijoy2004040@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all] (>=0.135.1,<0.136.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "project_1", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
